Added BSP textures + fixes & tweaks
