+++ src_c_api/wad.h
@@ -4,14 +4,12 @@
     #include <glm/glm.hpp>
 #endif
 
-static_assert(sizeof(unsigned int) == 4);
-static_assert(sizeof(unsigned char) == 1);
-
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
+    // 1 byte per channel, 4 channels (red, green, blue, alpha)
 #ifdef __cplusplus
     typedef glm::u8vec4 wad_rgba;
 #else
@@ -20,7 +18,6 @@ extern "C"
         unsigned char r, g, b, a;
     } wad_rgba;
 #endif
-    static_assert(sizeof(wad_rgba) == 4); // 1 byte per channel, 4 channels (red, green, blue, alpha)
 
     typedef struct {
         /// Name of the texture
